# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: Front Office API — league management, results, standings, and scheduling.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues/{leagueId}/settings:
    parameters:
      - $ref: '#/components/parameters/LeagueId'
    get:
      tags: [Leagues]
      operationId: getLeagueSettings
      summary: Get the active immutable settings version (league members)
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Active settings version, or a null data result when none exists yet.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LeagueSettingsVersion'
                  - $ref: '#/components/schemas/NoLeagueSettings'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Caller is not an active league member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: League or settings not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /leagues/{leagueId}/settings/versions:
    parameters:
      - $ref: '#/components/parameters/LeagueId'
    post:
      tags: [Leagues]
      operationId: createLeagueSettingsVersion
      summary: Append and activate a new settings version (commissioner only)
      description: Existing settings versions are never updated or deleted.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeagueSettingsInput'
      responses:
        '201':
          description: New immutable settings version created and activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueSettingsVersion'
        '400':
          description: Invalid settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Caller is not the commissioner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Settings conflict with active season occupancy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /leagues/{leagueId}/settings/history:
    parameters:
      - $ref: '#/components/parameters/LeagueId'
    get:
      tags: [Leagues]
      operationId: listLeagueSettingsHistory
      summary: List immutable settings history newest first (league members)
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Settings history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueSettingsHistory'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Caller is not an active league member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  parameters:
...
    LeagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 255
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status, trace_id]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: ["string", "null"]
        instance:
          type: ["string", "null"]
        trace_id:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemError'
    ProblemError:
      type: object
      properties:
        pointer:
          type: string
        message:
          type: string
...
    LeagueSettingsInput:
      type: object
      additionalProperties: false
      required:
        - platform
        - team_count
        - roster_source
        - schedule_format
        - schedule_settings
        - playoff_format
        - divisions
        - conferences
        - slider_presets
        - require_verified_identities
        - points_win
        - points_ot_loss
        - points_reg_loss
        - tiebreakers
        - use_league_defaults
        - change_summary
      properties:
        applied_template_id:
          type: ["string", "null"]
          enum: [balanced_standard, international_style, maximum_fighting_contact, null]
          description: Curated template explicitly applied before this version was saved.
        ea_league_id:
          type: ["string", "null"]
          maxLength: 100
        platform:
          type: string
          enum: [psn, xbox, both]
        team_count:
          type: integer
          minimum: 3
          maximum: 32
        roster_source:
          type: string
          enum: [manual, ea, csv_import]
        schedule_format:
          type: string
          enum: [round_robin, double_round_robin, custom]
        schedule_settings:
          type: object
          additionalProperties: true
        playoff_format:
          type: object
          additionalProperties: true
        salary_cap_cents:
          type: ["integer", "null"]
          minimum: 0
        roster_min:
          type: ["integer", "null"]
          minimum: 1
        roster_max:
          type: ["integer", "null"]
          minimum: 1
        divisions:
          type: array
          items:
            type: string
        conferences:
          type: array
          items:
            type: string
        rules_notes:
          type: ["string", "null"]
        slider_presets:
          type: object
          additionalProperties: true
        require_verified_identities:
          type: boolean
          default: false
          description: Require the selected saved console identity to be currently verified for signup and waitlist eligibility.
        points_win:
          type: integer
          minimum: 0
        points_ot_loss:
          type: integer
          minimum: 0
        points_reg_loss:
          type: integer
          minimum: 0
        tiebreakers:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
        use_league_defaults:
          type: boolean
          default: true
        apply_to_active_season:
          type: boolean
          default: false
        change_summary:
          type: string
          minLength: 1
          maxLength: 500
...
    LeagueSettingsVersion:
      allOf:
        - $ref: '#/components/schemas/LeagueSettingsInput'
        - type: object
          required: [id, league_id, version, changed_by, changed_at, is_active, can_manage]
          properties:
            id:
              type: string
            league_id:
              type: string
            version:
              type: integer
            changed_by:
              type: string
            changed_at:
              type: string
              format: date-time
            is_active:
              type: boolean
            can_manage:
              type: boolean
    LeagueSettingsHistory:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LeagueSettingsVersion'
    NoLeagueSettings:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: "null"
...